+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/conceptXY/LineDirectionForXY.kt
@@ -1,4 +1,4 @@
-package geometry
+package geometry.conceptXY
 
 import elements.LineDirection
