+++ Attt_LocalEngine_Kotlin/src/main/kotlin/geometry/conceptXY/NearestAreaScanWithXY.kt
@@ -1,4 +1,4 @@
-package geometry
+package geometry.conceptXY
 
 import elements.Coordinates
 import logic.GameField
@@ -6,69 +6,63 @@ import logic.OneMoveProcessing
 import logic.PlayerProvider
 import utilities.Log
 
-internal class NearestAreaScanWith2D(private val gameField: GameField) : OneMoveProcessing {
+internal class NearestAreaScanWithXY(private val gameField: GameField) : OneMoveProcessing {
 
     override fun getMaxLengthAchievedForThisMove(where: Coordinates): Int? {
-        if (where !is Coordinates2D) return null
+        if (where !is CoordinatesXY) return null
         return detectAllExistingLineDirectionsFromThePlacedMark(where)
-            .maxOfOrNull { twoAxisDirection ->
-                measureFullLengthForExistingLineFrom(where, twoAxisDirection)
+            .maxOfOrNull { lineDirection ->
+                measureFullLengthForExistingLineFrom(where, lineDirection)
             }
     }
 
-    override fun getCoordinatesFor(x: Int, y: Int, z: Int): Coordinates = Coordinates2D(x, y)
+    override fun getCoordinatesFor(x: Int, y: Int, z: Int): Coordinates = CoordinatesXY(x, y)
 
-    private fun detectAllExistingLineDirectionsFromThePlacedMark(fromWhere: Coordinates2D): List<LineDirectionFor2Axes> {
+    private fun detectAllExistingLineDirectionsFromThePlacedMark(fromWhere: CoordinatesXY): List<LineDirectionForXY> {
         val checkedMark = gameField.getCurrentMarkAt(fromWhere)
         if (checkedMark == null || checkedMark == PlayerProvider.None) {
             return emptyList() // preventing from doing detection calculations for initially wrong Player
         }
-        val allDirections = mutableListOf<LineDirectionFor2Axes>()
-        LineDirectionFor2Axes.getAllFromLoops().filter { !it.isNone() }.forEach { twoAxisDirection ->
-            val nextCoordinates = fromWhere.getNextInTheDirection(
-                twoAxisDirection.xAxisLD, twoAxisDirection.yAxisLD
-            )
+        val allDirections = mutableListOf<LineDirectionForXY>()
+        LineDirectionForXY.entries.filter { it != LineDirectionForXY.None }.forEach { lineDirection ->
+            val nextCoordinates = fromWhere.getNextInTheDirection(lineDirection)
             if (nextCoordinates.existsWithin(gameField.sideLength) &&
                 gameField.containsTheSameMark(checkedMark, nextCoordinates)
             ) {
-                allDirections.add(twoAxisDirection)
-                Log.pl("line exists in direction: $twoAxisDirection")
+                allDirections.add(lineDirection)
+                Log.pl("line exists in direction: $lineDirection")
             }
         }
         return allDirections // is empty if no lines ae found in all possible directions
     }
 
     private fun measureFullLengthForExistingLineFrom(
-        start: Coordinates2D, lineDirectionFor2Axes: LineDirectionFor2Axes
+        start: CoordinatesXY, lineDirectionForXY: LineDirectionForXY
     ): Int {
         // here we already have a detected line of 2 minimum dots, now let's measure its full potential length.
         // we also have a proven placed dot of the same player in the detected line direction.
         // so, we only have to inspect next potential dot of the same direction -> let's prepare the coordinates:
-        val checkedNearCoordinates = start.getTheNextSpaceFor(
-            lineDirectionFor2Axes.xAxisLD, lineDirectionFor2Axes.yAxisLD, gameField.sideLength
-        )
+        val checkedNearCoordinates = start.getTheNextSpaceFor(lineDirectionForXY, gameField.sideLength)
         var lineTotalLength = 0
-        if (checkedNearCoordinates is Coordinates2D) {
+        if (checkedNearCoordinates is CoordinatesXY) {
             lineTotalLength =
-                measureLineFrom(checkedNearCoordinates, lineDirectionFor2Axes, 2) +
-                        measureLineFrom(start, lineDirectionFor2Axes.opposite(), 0)
+                measureLineFrom(checkedNearCoordinates, lineDirectionForXY, 2) +
+                        measureLineFrom(start, lineDirectionForXY.opposite(), 0)
             Log.pl("makeNewMove: lineTotalLength = $lineTotalLength")
         } // else checkedNearCoordinates cannot be Border or anything else apart from Coordinates type
         return lineTotalLength
     }
 
-    private fun measureLineFrom(
-        givenMark: Coordinates2D, lineDirectionFor2Axes: LineDirectionFor2Axes, startingLength: Int
+    internal fun measureLineFrom(
+        givenMark: CoordinatesXY, lineDirectionForXY: LineDirectionForXY, startingLength: Int
     ): Int {
         Log.pl("measureLineFrom: given startingLength: $startingLength")
         Log.pl("measureLineFrom: given start coordinates: $givenMark")
         // firstly let's measure in the given direction and then in the opposite, also recursively
-        val nextMark = givenMark.getTheNextSpaceFor(
-            lineDirectionFor2Axes.xAxisLD, lineDirectionFor2Axes.yAxisLD, gameField.sideLength
-        )
+        val nextMark = givenMark.getTheNextSpaceFor(lineDirectionForXY, gameField.sideLength)
         Log.pl("measureLineFrom: detected next coordinates: $nextMark")
-        return if (nextMark is Coordinates2D && gameField.belongToTheSameRealPlayer(givenMark, nextMark)) {
-            measureLineFrom(nextMark, lineDirectionFor2Axes, startingLength + 1)
+        return if (nextMark is CoordinatesXY && gameField.belongToTheSameRealPlayer(givenMark, nextMark)) {
+            measureLineFrom(nextMark, lineDirectionForXY, startingLength + 1)
         } else {
             Log.pl("measureLineFrom: ELSE -> exit: $startingLength")
             startingLength
